[html-formatter] Add a test-cases event
We're working towards adding a progress bar. We figure that in order to
build a progress bar, we need to first know the total number of test
cases that Cucumber is about to execute. Instead of putting this in the
payload of another message, we chose to put it in its own.

pairing with @jbpros